scripts: render-ida-import-script: add doc
